Concept of the game
I have  listed the content of the game, of how it should be and added unique attributes to the enemy, so the game doesn't get boring for the player. I also wrote down a little bit of the story, which isn't really important. The Dialogues of the NPC's will be coming from another md-file, which contains all the dialogues of the Interactions of an Object.